backend/requirements.txt
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt

django-cors-headers==3.11.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
celery==5.2.7
redis==4.3.4
django-environ==0.9.0

requests==2.26.0
